+++ docs/source/api/randprocs/markov/continuous.rst
@@ -4,8 +4,3 @@ probnum.randprocs.markov.continuous
 .. automodapi:: probnum.randprocs.markov.continuous
     :no-heading:
     :headings: "*"
-
-.. toctree::
-    :hidden:
-
-    continuous/integrator
