--- docs/source/api/randprocs/markov/continuous/integrator.rst
@@ -1,6 +1,0 @@
-probnum.randprocs.markov.continuous.integrator
-**********************************************
-
-.. automodapi:: probnum.randprocs.markov.continuous.integrator
-    :no-heading:
-    :headings: "^"
