+++ docs/source/api/randprocs/markov/integrator.rst
@@ -1,0 +1,6 @@
+probnum.randprocs.markov.integrator
+-----------------------------------
+
+.. automodapi:: probnum.randprocs.markov.continuous.integrator
+    :no-heading:
+    :headings: "*"
