+++ src/probnum/problems/zoo/filtsmooth/_filtsmooth_problems.py
@@ -96,7 +96,7 @@ def car_tracking(
     state_dim = 2
     model_dim = state_dim * (model_nu + 1)
     measurement_dim = 2
-    dynamics_model = randprocs.markov.continuous.integrator.IntegratedWienerTransition(
+    dynamics_model = randprocs.markov.integrator.IntegratedWienerTransition(
         num_derivatives=model_nu,
         wiener_process_dimension=state_dim,
         forward_implementation=forward_implementation,
@@ -211,14 +211,12 @@ def ornstein_uhlenbeck(
         Cambridge University Press, 2019
     """
 
-    dynamics_model = (
-        randprocs.markov.continuous.integrator.IntegratedOrnsteinUhlenbeckTransition(
-            num_derivatives=0,
-            wiener_process_dimension=1,
-            driftspeed=driftspeed,
-            forward_implementation=forward_implementation,
-            backward_implementation=backward_implementation,
-        )
+    dynamics_model = randprocs.markov.integrator.IntegratedOrnsteinUhlenbeckTransition(
+        num_derivatives=0,
+        wiener_process_dimension=1,
+        driftspeed=driftspeed,
+        forward_implementation=forward_implementation,
+        backward_implementation=backward_implementation,
     )
     dynamics_model.dispmat *= process_diffusion
 
@@ -569,7 +567,7 @@ def logistic_ode(
 
     t0, tmax = timespan
     logistic_ivp = diffeq.logistic(t0=t0, tmax=tmax, y0=y0, params=params)
-    dynamics_model = randprocs.markov.continuous.integrator.IntegratedWienerTransition(
+    dynamics_model = randprocs.markov.integrator.IntegratedWienerTransition(
         num_derivatives=order,
         wiener_process_dimension=1,
         forward_implementation=forward_implementation,
