+++ src/probnum/randprocs/markov/integrator/_integrator.py
@@ -5,7 +5,7 @@ This is the sucessor of the former ODEPrior.
 
 import numpy as np
 
-from probnum.randprocs.markov.continuous.integrator import _preconditioner
+from probnum.randprocs.markov.integrator import _preconditioner
 
 __all__ = ["IntegratorTransition"]
