+++ tests/test_diffeq/test_odefiltsmooth/test_string_to_measmodel.py
@@ -22,7 +22,7 @@ def ivp():
 @pytest.fixture
 def prior(ivp):
     ode_dim = ivp.dimension
-    prior = randprocs.markov.continuous.integrator.IntegratedWienerTransition(
+    prior = randprocs.markov.integrator.IntegratedWienerTransition(
         num_derivatives=2, wiener_process_dimension=ode_dim
     )
     initrv = randvars.Normal(
